Lift/Lift: update to libtcg
Replaces PTC with libtcg. The dlopen interface of libtcg is used where a
single libtcg_load function needs to dlsym'd. libtcg_load returns a
struct of function pointers that will be passed around to all code that
needs to call libtcg. This makes the dependence on libtcg explicit, and
in the future, handling multiple open libtcg libraries will be trivial.